Develop (#10)
* update for September 2024

* September 2024 update

* Refactor bill calculation and output logic

This commit refactors the bill calculation and output logic in the main.py file. It introduces an argparse module to handle command-line arguments for the file path to the YAML data. The main function is updated to use the argparse module and create an instance of the BillCalculator class. The calculate and output_bill methods are called to calculate and output the bills using the TableOutputStrategy and CSVOutputStrategy. This refactoring improves the readability and maintainability of the code.

* Update .gitignore

disable a pylint rule

* Update bill_type.py

disable a pylint rule

* Update bill_type.py

pylint rule